Refactor deleting extra thumbnails in umrah package gallery middleware

+++ middlewares/umrah/packages/upload-umrah-package-gallery.js
@@ -36,9 +36,10 @@ module.exports =
         // check if umrah package extra thumbnails exists
         if (umrahPackage?.extraThumbnails?.length > 0) {
             // delete previous extra thumbnails
-            extraThumbnails.forEach((thumbnail) => {
-                fs.unlinkSync(path.join(__dirname, '../../../public/', thumbnail));
-            });
+            extraThumbnails.forEach(
+                (thumbnail) =>
+                    thumbnail && fs.unlinkSync(path.join(__dirname, '../../../public/', thumbnail))
+            );
         }
 
         // prepare file path
